machine_learning/inference_service/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
google-cloud-storage==2.19.0

joblib==1.4.2

numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
xgboost==3.0.2

# Data validation and settings management
pydantic==2.10.6
pydantic-settings==2.10.1
